+++ Sources/StreamChatUI/CommonViews/SwiftUIViewRepresentable.swift
@@ -26,6 +26,19 @@ public extension SwiftUIRepresentable where Self: UIView {
 }
 
 @available(iOS 13.0, *)
+public extension SwiftUIRepresentable where Self: UIViewController {
+    /// Creates `SwiftUIViewControllerRepresentable` instance wrapping the current type that can be used in your SwiftUI view
+    /// - Parameters:
+    ///     - content: Content of the view controller. Its value is automatically updated when it's changed
+    static func asView(_ content: ViewContent) -> SwiftUIViewControllerRepresentable<Self> {
+        SwiftUIViewControllerRepresentable(
+            viewController: self,
+            content: content
+        )
+    }
+}
+
+@available(iOS 13.0, *)
 /// A concrete type that wraps a view conforming to `SwiftUIRepresentable` and enables using it in SwiftUI via `UIViewRepresentable`
 public struct SwiftUIViewRepresentable<View: UIView & SwiftUIRepresentable>: UIViewRepresentable {
     private let view: View.Type
@@ -46,4 +59,30 @@ public struct SwiftUIViewRepresentable<View: UIView & SwiftUIRepresentable>: UIV
     public func updateUIView(_ uiView: View, context: Context) {
         uiView.content = content
     }
+}
+
+@available(iOS 13.0, *)
+/// A concrete type that wraps a view conforming to `SwiftUIRepresentable` and enables using it in SwiftUI via `UIViewControllerRepresentable`
+public struct SwiftUIViewControllerRepresentable<
+    ViewController: UIViewController &
+        SwiftUIRepresentable
+>: UIViewControllerRepresentable {
+    private let viewController: ViewController.Type
+    private let content: ViewController.ViewContent
+    
+    init(
+        viewController: ViewController.Type,
+        content: ViewController.ViewContent
+    ) {
+        self.viewController = viewController
+        self.content = content
+    }
+    
+    public func makeUIViewController(context: Context) -> ViewController {
+        let controller = ViewController()
+        controller.content = content
+        return controller
+    }
+    
+    public func updateUIViewController(_ uiViewController: ViewController, context: Context) {}
 }